StringDataLayer->GetRenderLines() bug fixed.

+++ WpfHexaEditor.Shared/Core/HexBrushPosition.cs
@@ -7,7 +7,7 @@ using WpfHexaEditor.Core.Interfaces;
 namespace WpfHexaEditor.Core
 {
     class HexBrushPosition : IHexBrushPosition {
-        public long Position { get; set; }
+        public int Position { get; set; }
 
         public Brush FirstCharBrush { get; set; }
